Homework_05 + added endpoints/swagger discription + postman

+++ src/main/java/de/ait/training/model/Car.java
@@ -10,19 +10,21 @@ import jakarta.persistence.Table;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
-// либо просто добавляем аннотацию @data
+// lombok
+// Equivalent to @Getter @Setter @RequiredArgsConstructor @ToString @EqualsAndHashCode.
 @Data
 @NoArgsConstructor
+// Таблица в БД будет представлена через клас Car
 @Entity
 @Table(name = "cars")
 
 
 public class Car {
+    // уникальный идентификатор для каждой транзакции
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
+    // перед каждым сохранением в БД проверяет на null
     @Column(nullable = false)
     private String color;
 
@@ -30,9 +32,9 @@ public class Car {
     private String model;
 
     @Column(nullable = false)
-    private int price;
+    private double price;
 
-    public Car(String color, String model, int price) {
+    public Car(String color, String model, double price) {
         this.color = color;
         this.model = model;
         this.price = price;
